pyproject.toml
[project]
name = "natural-sql"
version = "0.1.0"
description = "Natural language to SQL query converter with CLI and web interface"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "openai>=1.0.0",
    "mysql-connector-python>=8.2.0",
    "python-dotenv>=1.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "tabulate>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0"
]

[project.scripts]
natural-sql = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
